+++ src/packages/extensions/webpack.prod.config.js
@@ -5,13 +5,13 @@ const commonConfig = require("../common.webpack.prod.config");
 
 const config = {
   entry: {
-    "excalidraw-plugins.production.min": "./index.ts",
+    "excalidraw-extensions.production.min": "./index.ts",
   },
   output: {
     path: path.resolve(__dirname, "dist"),
-    library: "ExcalidrawPluginsLib",
-    chunkFilename: "excalidraw-plugins-assets/[name]-[contenthash].js",
-    assetModuleFilename: "excalidraw-plugins-assets/[name][ext]",
+    library: "ExcalidrawExtensionsLib",
+    chunkFilename: "excalidraw-extensions-assets/[name]-[contenthash].js",
+    assetModuleFilename: "excalidraw-extensions-assets/[name][ext]",
   },
 };
 module.exports = merge(commonConfig, config);